add todo, option seems to not be working :(
first attempt at getting a proper wind chart.
Chart should ideally be getting data along the vertical axis and be centered around the average direction for the last N minutes with the connecting line for two data points always following the shortest route (i.e. not from 10 to 350 but rather from 350 to 10 via 0).